requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
proglog==0.1.10
python-vlc==3.0.20123
requests==2.31.0
tqdm==4.66.2
urllib3==2.2.0
